+++ NEWS.md
@@ -1,3 +1,6 @@
+# idaifieldR 0.2.3 _2023-02-18_
+* add language list lookup preparation
+
 # idaifieldR 0.2.2 _2023-02-18_
 * config specific names now removed everywhere
 * fixed problem with if-condition when config exists